custom_addons/gchat_integration/listener/requirements.txt
google-cloud-pubsub>=2.18.0
google-auth>=2.17.0
requests>=2.28.0 